+++ docs/CHANGES.rst
@@ -1,8 +1,8 @@
 Changelog
 =========
 
-1.3.4 (unreleased)
-------------------
+1.4 (2015-09-04)
+----------------
 
 - plone.api.content.delete: add option check_linkintegrity. If True raise
   exception if deleting would result in broken links.